Move a package.json logic to a function in script-util.
